Extract operators
