+++ apps/gears/route-story/common/src/index.ts
@@ -1,4 +1,5 @@
 export * from './route-story-gear';
+export * from './file-operator';
 export * from './model';
 export * from './layers';
 export * from './tinkers';
