+++ apps/gears/route-story/mobile/src/RouteStoryFileInput.tsx
@@ -3,12 +3,11 @@ import { View, } from "react-native";
 import { DocumentPickerResponse, types } from "@react-native-documents/picker";
 import RNFS from 'react-native-fs';
 import { RouteFileInputProps } from "@the-dead-planet/nav-gauge-gears-route-story-common";
-import { FileInputStatus, FileInput } from "@mobile-ui";
-import { parsers, useStateWarden, useSubjectState } from "@apparatus";
+import { FileInputStatus, FileInput, MobileMap } from "@mobile-ui";
+import { parsers, useSubjectState } from "@apparatus";
 import { useImageReader } from "./images/useImageReader";
 
-export const RouteStoryFileInput: FC<RouteFileInputProps> = ({ data$, images$, fileOperator }) => {
-    const { signaliumBureau } = useStateWarden();
+export const RouteStoryFileInput: FC<RouteFileInputProps<MobileMap>> = ({ data$, images$, fileOperator }) => {
     const [{ geojson, routeName, error }] = useSubjectState(data$);
     const [isLoading, setIsLoading] = useSubjectState(fileOperator.isLoading$);
     const readImage = useImageReader(fileOperator, images$)
@@ -16,7 +15,6 @@ export const RouteStoryFileInput: FC<RouteFileInputProps> = ({ data$, images$, f
     const handleUpload = async (files: DocumentPickerResponse[]) => {
         fileOperator.uploadFile<DocumentPickerResponse>(
             files,
-            signaliumBureau,
             (file) => RNFS.readFile(file.uri, 'utf8'),
             readImage,
         );
@@ -32,7 +30,7 @@ export const RouteStoryFileInput: FC<RouteFileInputProps> = ({ data$, images$, f
                 allowMultiSelection
                 onIsLoadingChange={setIsLoading}
                 onUpload={handleUpload}
-                onError={(err) => fileOperator.onError(err, signaliumBureau)}
+                onError={fileOperator.onError}
             />
             <FileInputStatus
                 isLoading={isLoading}
